+++ src/Sdk.php
@@ -6,7 +6,7 @@ use MercadoPago\PP\Sdk\Common\AbstractEntity;
 use MercadoPago\PP\Sdk\Common\Config;
 use MercadoPago\PP\Sdk\Common\Constants;
 use MercadoPago\PP\Sdk\Common\Manager;
-use MercadoPago\PP\Sdk\Entity\Monitoring\MelidataError;
+use MercadoPago\PP\Sdk\Entity\Monitoring\DatadogEvent;
 use MercadoPago\PP\Sdk\Entity\Notification\Notification;
 use MercadoPago\PP\Sdk\Entity\Payment\Multipayment;
 use MercadoPago\PP\Sdk\Entity\Payment\MultipaymentV2;
@@ -136,11 +136,11 @@ class Sdk
     }
 
     /**
-     * @return MelidataError
+     * @return DatadogEvent
      */
-    public function getMelidataErrorInstance()
+    public function getDatadogEventInstance()
     {
-        return $this->getEntityInstance('MercadoPago\PP\Sdk\Entity\Monitoring\MelidataError', Constants::BASEURL_MP);
+        return $this->getEntityInstance('MercadoPago\PP\Sdk\Entity\Monitoring\DatadogEvent', Constants::BASEURL_MP);
     }
 
     /**