+++ tests/unit/Mock/DatadogEventMock.php
@@ -3,23 +3,20 @@
 namespace MercadoPago\PP\Sdk\Tests\Unit\Mock;
 
 /**
- * Class MelidataErrorMock
+ * Class DatadogEventMock
  */
-class MelidataErrorMock
+class DatadogEventMock
 {
 
-  const COMPLETE_MELIDATA_ERROR = [
-    "name" => "error name",
+  const COMPLETE_DATADOG_EVENT = [
+    "value" => "success",
     "message" => "error message",
-    "target" => "error target",
-    "plugin" => [
-      "version" => "123",
-    ],
+    "plugin_version" => "1.2.3",
     "platform" => [
       "name" => "ppcoreinternal",
       "version" => "1.2.3",
       "uri" => "/platform_uri",
-      "location" => "platform_location",
+      "url" => "https://platform_url...",
     ],
     "details" => [
       "payment_id" => "123456",